requirements.txt
Django==1.10.7
django-crispy-forms==1.6.1
django-markdown-deux==1.0.5
django-pagedown==0.1.3
djangorestframework==3.6.3
markdown2==2.3.4
olefile==0.44
Pillow==4.2.0
pkg-resources==0.0.0
